wasm backend compiles but does nothing
