+++ cpu_kernel_raytracer/Cargo.toml
@@ -6,6 +6,7 @@ edition = "2018"
 
 [features]
 cuda = ["raytracer_lib_no_std/cuda", "math/cuda"]
+wasm = ["raytracer_lib_no_std/wasm", "math/wasm"]
 
 [dependencies]
 raytracer_lib_no_std = {path = "../raytracer_lib_no_std" }
