+++ math/Cargo.toml
@@ -8,6 +8,7 @@ edition = "2018"
 [features]
 cuda = [ "rustacuda_core", "rustacuda_derive" ]
 use_serde = [ "serde", "serde_json" , "serde_derive"]
+wasm = [ "libm" ]
 
 [dependencies]
 rustacuda_core = { path = "../../RustaCUDA/rustacuda_core/" , optional = true}
@@ -16,3 +17,5 @@ rustacuda_derive = { path = "../../RustaCUDA/rustacuda_derive/" , optional = tru
 serde = { version = "*", features = ["derive"], optional = true }
 serde_json ={  version = "*", optional = true }
 serde_derive ={  version = "*", optional = true }
+
+libm = { version="*", optional = true}
