+++ raytracer/Cargo.toml
@@ -7,7 +7,7 @@ edition = "2018"
 [features]
 cuda = ["rustacuda", "rustacuda_derive", "rustacuda_core", "cpu_kernel_raytracer/cuda"]
 use_serde = [ "raytracer_lib_std/use_serde", "raytracer_lib_no_std/use_serde" ]
-
+wasm = ["cpu_kernel_raytracer/wasm"]
 
 [dependencies]
 raytracer_lib_std = {path =  "../raytracer_lib_std" }
