+++ raytracer/src/backend/backend_cuda.rs
@@ -129,10 +129,6 @@ impl Backend for BackendCuda {
         }
         Ok(c)
     }
-
-    fn render_world_multi_core(&self, world: &mut World, c: &Camera) -> Result<Canvas, Box<dyn Error>> {
-        self.render_world(world, c)
-    }
 }
 
 impl BackendCuda {
