+++ raytracer_challenge_reference_impl/src/basics/mod.rs
@@ -1,12 +1,10 @@
 pub mod camera;
-pub mod canvas;
 pub mod color;
 pub mod intersection;
 pub mod precomputed_component;
 pub mod ray;
 
 pub use self::camera::*;
-pub use self::canvas::*;
 pub use self::color::*;
 pub use self::intersection::*;
 pub use self::precomputed_component::*;
