+++ raytracer_lib_no_std/src/basics/mod.rs
@@ -2,8 +2,10 @@ pub use self::camera::*;
 pub use self::color::*;
 pub use self::precomputed_component::*;
 pub use self::ray::*;
+pub use self::pixel::*;
 
 pub mod camera;
 pub mod color;
 pub mod precomputed_component;
 pub mod ray;
+pub mod pixel;
