+++ raytracer_lib_no_std/src/basics/pixel.rs
@@ -1,9 +1,7 @@
 #[cfg(feature = "use_serde")]
 use serde::{Deserialize, Serialize};
+use crate::{Color, BLACK};
 
-use raytracer_lib_no_std::{BLACK, Color};
-
-pub type PixelVec = Vec<Pixel>;
 
 #[derive(Clone, Debug)]
 #[cfg_attr(feature = "use_serde", derive(Deserialize, Serialize))]
