+++ raytracer_lib_std/Cargo.toml
@@ -6,16 +6,14 @@ edition = "2018"
 
 [features]
 use_serde = [ "serde", "serde_json" , "serde_derive", "math/use_serde", "math/use_serde", "raytracer_lib_no_std/use_serde" ]
+wasm = ["math/wasm", "raytracer_lib_no_std/wasm"]
 
 [dependencies]
 raytracer_lib_no_std = {path =  "../raytracer_lib_no_std"}
-image = "*"
 math = { path = "../math/" }
 
 serde = { version = "*", features = ["derive"], optional = true }
 serde_json ={  version = "*", optional = true }
 serde_derive ={  version = "*", optional = true }
 
-
-
-
+image = "*"
