--- raytracer_lib_std/src/canvas/mod.rs
@@ -1,5 +1,0 @@
-pub use self::canvas::*;
-pub use self::pixel::*;
-
-pub mod canvas;
-pub mod pixel;
