--- raytracer_lib_std/src/scene/mod.rs
@@ -1,3 +1,0 @@
-pub use self::scene::*;
-
-pub mod scene;
