--- raytracer_lib_std/src/world/mod.rs
@@ -1,3 +1,0 @@
-pub use self::world::*;
-
-pub mod world;
